more bbs

+++ Core/ORMModels/BollingerBand.cs
@@ -6,13 +6,13 @@ namespace Core.ORMModels
     {
         public string Symbol { get; set; }
         public DateTime Date { get; set; }
-        public decimal Open { get; set; }
-        public decimal High { get; set; }
-        public decimal Low { get; set; }
-        public decimal Close { get; set; }
-        public decimal SMA20 { get; set; }
-        public decimal UpperBand { get; set; }
-        public decimal LowerBand { get; set; }
-        public decimal BandRatio { get; set; }
+        public double Open { get; set; }
+        public double High { get; set; }
+        public double Low { get; set; }
+        public double Close { get; set; }
+        public double SMA20 { get; set; }
+        public double UpperBand { get; set; }
+        public double LowerBand { get; set; }
+        public double BandRatio { get; set; }
     }
 }
